src/voicelive/requirements.txt
aiohttp==3.12.13
aiohttp-sse==2.2.0
azure-identity==1.23.0
azure.cosmos==4.9.0
gunicorn==23.0.0
openai==1.90.0
python-dotenv==1.1.0